Haha :)) Loss Overflow
I have the loss overflow when trying to calculate the loss_style in the
Perceptual Loss algorithm. The reason of that is because of float16
blocks the range of calculation to 65,504. However, the gram matrix need
to divide (c * h * w) which is around 4,194,304 if c = 64, h = 256, w =
256 (at the relu_1_2 layer). Therefore, that values become Nan or Inf.
To counter that problem, we need to switch back to float32 of each
tensors before feeding to VGG Model