Changing the Architecture of the Attention Block
I have detected the OOM problem. It comes from the fact that the
Attention Block incorporates all the pixels together. For example, if we
have the image with dimension 128x128, the attenion map will results in
a (128^2 x 128^2) table, overloading the memory. Thus, I add a section
dividing the image into small patches. With this patch, we can perform
global attention (getting the global features) while preventing exceeded
usage of GPU memory